move plugin to flux controller to flux pkg
