+++ internal/controller/flux/util.go
@@ -12,6 +12,11 @@ import (
 	sourcecontroller "github.com/fluxcd/source-controller/api/v1"
 )
 
+const (
+	deliveryToolLabel = "greenhouse.sap/deployment-tool"
+	deliveryToolFlux  = "flux"
+)
+
 func convertName(repoName string) (convertedName, repoType string) {
 	repoType = sourcecontroller.HelmRepositoryTypeDefault
 	// set the helm repository type to OCI if the repo name starts with oci://
@@ -36,13 +41,13 @@ func convertName(repoName string) (convertedName, repoType string) {
 	return convertedName, repoType
 }
 
-func FindHelmRepositoryByUrl(ctx context.Context, k8sClient client.Client, nS, url string) *sourcecontroller.HelmRepository {
+func findHelmRepositoryByUrl(ctx context.Context, k8sClient client.Client, nS, url string) *sourcecontroller.HelmRepository {
 	helmRepositoryList := new(sourcecontroller.HelmRepositoryList)
 	if err := k8sClient.List(ctx, helmRepositoryList); err != nil {
 		return nil
 	}
 	for _, helmRepository := range helmRepositoryList.Items {
-		if helmRepository.Spec.URL == url {
+		if helmRepository.Namespace == nS && helmRepository.Spec.URL == url {
 			return &helmRepository
 		}
 	}
