+++ internal/controller/plugin/plugin_controller.go
@@ -113,7 +113,7 @@ func (r *PluginReconciler) setConditions() lifecycle.Conditioner {
 			return
 		}
 
-		readyCondition := computeReadyCondition(plugin.Status.StatusConditions)
+		readyCondition := ComputeReadyCondition(plugin.Status.StatusConditions)
 		plugin.SetCondition(readyCondition)
 	}
 }
@@ -146,7 +146,7 @@ func (r *PluginReconciler) EnsureDeleted(ctx context.Context, resource lifecycle
 func (r *PluginReconciler) EnsureCreated(ctx context.Context, resource lifecycle.RuntimeObject) (ctrl.Result, lifecycle.ReconcileResult, error) {
 	plugin := resource.(*greenhousev1alpha1.Plugin) //nolint:errcheck
 
-	initPluginStatus(plugin)
+	InitPluginStatus(plugin)
 
 	restClientGetter, err := initClientGetter(ctx, r.Client, r.kubeClientOpts, *plugin)
 	if err != nil {
@@ -371,18 +371,18 @@ func (r *PluginReconciler) enqueueAllPluginsForCluster(ctx context.Context, o cl
 		FieldSelector: fields.OneTermEqualSelector(greenhouseapis.PluginClusterNameField, o.GetName()),
 		Namespace:     o.GetNamespace(),
 	}
-	return listPluginsAsReconcileRequests(ctx, r.Client, listOpts)
+	return ListPluginsAsReconcileRequests(ctx, r.Client, listOpts)
 }
 
 func (r *PluginReconciler) enqueueAllPluginsInNamespace(ctx context.Context, o client.Object) []ctrl.Request {
-	return listPluginsAsReconcileRequests(ctx, r.Client, client.InNamespace(o.GetNamespace()))
+	return ListPluginsAsReconcileRequests(ctx, r.Client, client.InNamespace(o.GetNamespace()))
 }
 
 func (r *PluginReconciler) enqueueAllPluginsForPluginDefinition(ctx context.Context, o client.Object) []ctrl.Request {
-	return listPluginsAsReconcileRequests(ctx, r.Client, client.MatchingLabels{greenhouseapis.LabelKeyPluginDefinition: o.GetName()})
+	return ListPluginsAsReconcileRequests(ctx, r.Client, client.MatchingLabels{greenhouseapis.LabelKeyPluginDefinition: o.GetName()})
 }
 
-func listPluginsAsReconcileRequests(ctx context.Context, c client.Client, listOpts ...client.ListOption) []ctrl.Request {
+func ListPluginsAsReconcileRequests(ctx context.Context, c client.Client, listOpts ...client.ListOption) []ctrl.Request {
 	var pluginList = new(greenhousev1alpha1.PluginList)
 	if err := c.List(ctx, pluginList, listOpts...); err != nil {
 		return nil
