+++ internal/controller/plugin/util.go
@@ -28,11 +28,6 @@ import (
 	"github.com/cloudoperators/greenhouse/internal/lifecycle"
 )
 
-const (
-	deliveryToolLabel = "greenhouse.sap/deployment-tool"
-	deliveryToolFlux  = "flux"
-)
-
 // exposedConditions are the conditions that are exposed in the StatusConditions of the Plugin.
 var exposedConditions = []greenhousemetav1alpha1.ConditionType{
 	greenhousemetav1alpha1.ReadyCondition,
@@ -48,8 +43,8 @@ type reconcileResult struct {
 	requeueAfter time.Duration
 }
 
-// initPluginStatus initializes all empty Plugin Conditions to "unknown"
-func initPluginStatus(plugin *greenhousev1alpha1.Plugin) greenhousev1alpha1.PluginStatus { //nolint:unparam
+// InitPluginStatus initializes all empty Plugin Conditions to "unknown"
+func InitPluginStatus(plugin *greenhousev1alpha1.Plugin) greenhousev1alpha1.PluginStatus { //nolint:unparam
 	for _, t := range exposedConditions {
 		if plugin.Status.GetConditionByType(t) == nil {
 			plugin.SetCondition(greenhousemetav1alpha1.UnknownCondition(t, "", ""))
@@ -243,8 +238,8 @@ func setWorkloadMetrics(plugin *greenhousev1alpha1.Plugin, status float64) {
 	workloadStatus.WithLabelValues(plugin.GetNamespace(), plugin.Name, plugin.Spec.PluginDefinition).Set(status)
 }
 
-// computeReadyCondition computes the ReadyCondition for the Plugin based on various status conditions
-func computeReadyCondition(
+// ComputeReadyCondition computes the ReadyCondition for the Plugin based on various status conditions
+func ComputeReadyCondition(
 	conditions greenhousemetav1alpha1.StatusConditions,
 ) (readyCondition greenhousemetav1alpha1.Condition) {
